100% Influence Fix (#67)
- Updated project file to version 4.1.5 with a fix for "100 Percent Influence".
- Changed influence properties from double to decimal in multiple files to prevent errors when a faction has 100% influence.
- Modified database parameter handling to accommodate decimal types.
- Adjusted parsing methods to ensure consistency across the application.
- Updated installer configuration to reflect the new version.

+++ FalconsFactionMonitor/FalconsFactionMonitor/Models/FactionDetail.cs
@@ -4,8 +4,8 @@
     {
         public string SystemName { get; set; }
         public string FactionName { get; set; }
-        public double InfluencePercent { get; set; }
-        public double Difference { get; set; }
+        public decimal InfluencePercent { get; set; }
+        public decimal Difference { get; set; }
         public bool IsPlayer { get; set; }
         public string LastUpdated { get; set; }
     }
